requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-WTF>=1.2.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
fpdf2>=2.8.0
requests>=2.31.0
pypdfium2>=4.0.0
